requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic==2.7.1
